+++ ergo_std/src/match.rs
@@ -7,9 +7,15 @@ use ergo_runtime::{metadata::Source, traits, types, Context, Value};
 ///
 /// Arguments: `:value (Array :bindings)`
 ///
-/// Returns the value resulting from the first value in `bindings` that doesn't produce a pattern
-/// error when bound with `value`.
-pub async fn function(mut value: _, bindings: types::Array) -> Value {
+/// Keyed Arguments:
+/// * `:fallback` - A fallback value to bind the match value to if no other bindings match. This
+/// differs from adding a final `:a -> ...` case in error handling: if binding to the fallback
+/// value results in an error it is returned, whereas if binding to an equivalent final case
+/// results in an error it is interpreted as not matching (which results in a generic error).
+///
+/// Returns the value resulting from the first value in `bindings` that doesn't produce an error
+/// when bound with `value`.
+pub async fn function(mut value: _, bindings: types::Array, (fallback): [_]) -> Value {
     let bindings_source = Source::get(&bindings);
     let bindings = bindings.to_owned().0;
 
@@ -36,6 +42,14 @@ pub async fn function(mut value: _, bindings: types::Array) -> Value {
                 }
             }
 
+            // If a fallback is provided, bind it prior to inspecting the argument.
+            if let Some(fallback) = fallback {
+                return Ok(ergo_runtime::try_value!(
+                    traits::bind(fallback, value).await
+                ));
+            }
+
+            // If the value is an error and no bindings match, return it directly.
             let err = match value.as_type::<types::Error>() {
                 Ok(e) => e.to_owned(),
                 Err(v) => Source::get(&v)
@@ -76,6 +90,11 @@ mod test {
 
         fn match_case_body_bind_failure(t) {
             t.assert_fail("self:match 1 [1 -> {fn x -> () |> y}, 1 -> 1]");
+        }
+
+        fn match_fallback(t) {
+            t.assert_content_eq("a = _ -> !self:type:Error: doh; self:type:Error _ = self:match (fallback=:a) hi []; ()", "()");
+            t.assert_fail("a = _ -> !self:type:Error: doh; self:match hi [:a]; ()");
         }
     }
 }